Refactor y comienzo de eventos en menú lateral
