+++ src/components/MapLeaflet/MapViewLeaf.jsx
@@ -6,8 +6,8 @@ import { MarkersPlaces } from './MarkersPlaces'
 import { Routing } from './Routing'
 
 import 'leaflet/dist/leaflet.css'
-import { ModalDetalles } from '../modalDetalles'
-import { ModalEventos } from '../modalEventos'
+import { ModalDetalles } from '../Modals/modalDetalles'
+import { ModalEventos } from '../Modals/modalEventos'
 import { Modal } from 'react-bootstrap'
 import { MapContainer, TileLayer, ZoomControl } from 'react-leaflet'
 import 'leaflet-routing-machine'
